+++ packages/vendure-plugin-accept-blue/src/index.ts
@@ -2,3 +2,7 @@ export * from '../../util/src/subscription/default-subscription-strategy';
 export * from '../../util/src/subscription/subscription-strategy';
 export * from '../../util/src/subscription/subscription-helper';
 export * from './accept-blue-plugin';
+export * from './api/accept-blue-service';
+export * from './api/accept-blue-client';
+export * from './api/accept-blue-handler';
+export * from './api/accept-blue-common-resolvers';