+++ packages/vendure-plugin-accept-blue/src/types.ts
@@ -1,24 +1,6 @@
-interface CustomFields {
-  custom1: string;
-  custom2: string;
-  custom3: string;
-  custom4: string;
-  custom5: string;
-  custom6: string;
-  custom7: string;
-  custom8: string;
-  custom9: string;
-  custom10: string;
-  custom11: string;
-  custom12: string;
-  custom13: string;
-  custom14: string;
-  custom15: string;
-  custom16: string;
-  custom17: string;
-  custom18: string;
-  custom19: string;
-  custom20: string;
+import type { Request } from 'express';
+export interface CustomFields {
+  custom1?: string;
 }
 
 enum AcceptBlueAVSResultCode {
@@ -62,6 +44,7 @@ interface AcceptBlueTerminal {
   print_capability: boolean;
 }
 
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 interface TransactionFlags {
   allow_partial_approval: boolean;
   is_recurring: boolean;
@@ -78,6 +61,16 @@ interface TransactionDetails {
   terminal: string;
   client_ip: string;
   signature: string;
+  invoice_number: string;
+  po_number: string;
+  order_number: string;
+  batch_id: number;
+  source: string;
+  terminal_name: string;
+  username: string;
+  type: 'charge' | 'credit';
+  reference_number: number;
+  schedule_id: number;
 }
 
 export interface AcceptBlueAmountInput {
@@ -102,6 +95,7 @@ interface TransactionCustomer {
   customer_id: number;
 }
 
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 interface ChargeCustomer extends TransactionCustomer {
   send_receipt: boolean;
 }
@@ -334,6 +328,7 @@ export interface AcceptBlueTransaction {
   status_code: 'A' | 'P' | 'D' | 'E';
   error_message: string;
   error_code: string;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/no-redundant-type-constituents
   error_details: string | any;
   reference_number: number;
 }
@@ -349,13 +344,13 @@ export interface AcceptBlueChargeTransaction extends AcceptBlueTransaction {
   last_4: string;
   card_ref: string | null;
   bin_type?: AcceptBlueBinType;
-  transaction: CardTransaction | null;
+  transaction?: CardTransaction;
 }
 
 export interface AcceptBlueEvent {
   type: 'succeeded' | 'updated' | 'declined' | 'error' | 'status';
   subType: string;
-  event: 'transaction';
+  event: 'transaction' | 'batch';
   id: string;
   timestamp: string;
   data: AcceptBlueChargeTransaction;
@@ -410,4 +405,22 @@ export interface BatchEvent {
   id: string;
   timestamp: string;
   data: BatchObject;
+}
+
+export interface AcceptBlueWebhookInput {
+  webhook_url: string;
+  description: string;
+  active: true;
+}
+
+export interface AcceptBlueWebhook {
+  id: number;
+  signature: string;
+  webhook_url: string;
+  description: string;
+  active: boolean;
+}
+
+export interface RequestWithRawBody extends Request {
+  rawBody: Buffer;
 }