+++ packages/vendure-plugin-accept-blue/test/dev-server.ts
@@ -17,7 +17,7 @@ import {
 import readline from 'readline';
 import { AcceptBluePlugin } from '../src';
 import { acceptBluePaymentHandler } from '../src/api/accept-blue-handler';
-import { AcceptBlueTestCheckoutPlugin } from './accept-blue-test-checkout.plugin';
+import { AcceptBlueTestCheckoutPlugin } from './helpers/accept-blue-test-checkout.plugin';
 import {
   ADD_ITEM_TO_ORDER,
   ADD_PAYMENT_TO_ORDER,
@@ -26,9 +26,10 @@ import {
   REFUND_TRANSACTION,
   SET_SHIPPING_METHOD,
   TRANSITION_ORDER_TO,
-} from './helpers';
+} from './helpers/graphql-helpers';
 import { NoncePaymentMethodInput } from '../src/types';
 import { add } from 'date-fns';
+import { TestSubscriptionStrategy } from './helpers/test-subscription-strategy';
 
 /**
  * Ensure you have a .env in the plugin root directory with the variable ACCEPT_BLUE_TOKENIZATION_SOURCE_KEY=pk-abc123
@@ -55,7 +56,7 @@ import { add } from 'date-fns';
   const config: Required<VendureConfig> = mergeConfig(testConfig, {
     logger: new DefaultLogger({ level: LogLevel.Debug }),
     dbConnectionOptions: {
-      autoSave: true, // Uncomment this line to persist the database between restarts
+      // autoSave: true, // Uncomment this line to persist the database between restarts
     },
     authOptions: {
       cookieOptions: {
@@ -68,7 +69,13 @@ import { add } from 'date-fns';
     },
     plugins: [
       AcceptBlueTestCheckoutPlugin,
-      AcceptBluePlugin.init({}),
+      AcceptBluePlugin.init({
+        // vendureHost: process.env.VENDURE_HOST as string,
+        // Our temp webhook to catch the webhook from Accept Blue. View on: https://webhook.site/#!/view/cdef50e0-0e6d-4e23-a4b1-6ffc9ca89df8/2077d0c4-7cfb-4c81-b966-370ba5a44d7e/1
+        vendureHost:
+          'https://webhook.site/cdef50e0-0e6d-4e23-a4b1-6ffc9ca89df8',
+        subscriptionStrategy: new TestSubscriptionStrategy(),
+      }),
       DefaultSearchPlugin,
       AdminUiPlugin.init({
         port: 3002,
@@ -94,13 +101,14 @@ import { add } from 'date-fns';
   await adminClient.asSuperAdmin();
   await adminClient.query(CREATE_PAYMENT_METHOD, {
     input: {
-      code: 'accept-blue-credit-card',
+      code: 'accept-blue',
       enabled: true,
       handler: {
         code: acceptBluePaymentHandler.code,
         arguments: [
           { name: 'apiKey', value: process.env.API_KEY },
           { name: 'pin', value: process.env.PIN },
+          { name: 'testMode', value: 'true' },
           {
             name: 'tokenizationSourceKey',
             value: process.env.ACCEPT_BLUE_TOKENIZATION_SOURCE_KEY ?? null,
@@ -116,6 +124,7 @@ import { add } from 'date-fns';
     },
   });
   console.log(`Created paymentMethod`);
+
   await shopClient.asUserWithCredentials('hayden.zieme12@hotmail.com', 'test');
   await shopClient.query(ADD_ITEM_TO_ORDER, {
     productVariantId: '3',
@@ -147,7 +156,7 @@ import { add } from 'date-fns';
   try {
     const { addPaymentToOrder } = await shopClient.query(ADD_PAYMENT_TO_ORDER, {
       input: {
-        method: 'accept-blue-credit-card',
+        method: 'accept-blue',
         // metadata,
         metadata: { paymentMethodId: 14556 }, // Use a saved payment method
       },
@@ -157,13 +166,13 @@ import { add } from 'date-fns';
     );
 
     // Attempt a refund
-    const { refundAcceptBlueTransaction } = await shopClient.query(
-      REFUND_TRANSACTION,
-      {
-        transactionId: 354653,
-      }
-    );
-    console.log(`Refunded transaction: ${refundAcceptBlueTransaction}`);
+    // const { refundAcceptBlueTransaction } = await shopClient.query(
+    //   REFUND_TRANSACTION,
+    //   {
+    //     transactionId: 354653,
+    //   }
+    // );
+    // console.log(`Refunded transaction: ${refundAcceptBlueTransaction}`);
   } catch (e) {
     // Catch to prevent server from terminating
     console.error(e);